+++ RayTracer/Polygonal.h
@@ -15,12 +15,12 @@ private:
 public:
 	Polygonal(const std::vector<Vertice>& vertices);
 
-	virtual bool hit(const Ray& ray, float t0, float t1, float& t);
-	virtual Point3F hit_normal(const Ray& ray, float t) const;
+	virtual bool hit(const Ray& ray, double t0, double t1, double& t);
+	virtual Point3F hit_normal(const Ray& ray, double t) const;
 
 	Point3F normal() const;
 	const std::vector<Vertice>& get_vertices() const { return vertices_; }
-	virtual QColor read_texture(const Ray& ray, float t) const;
+	virtual QColor read_texture(const Ray& ray, double t) const;
 private:
 	bool is_coplanar_() const;
 	bool is_convex_() const;
