+++ RayTracer/Sphere.h
@@ -11,12 +11,12 @@ private:
 	float radius_;
 
 public:
-	Sphere(const Vertice& center, float radius);
+	Sphere(const Vertice& center, double radius);
 
-	virtual bool hit(const Ray& ray, float t0, float t1, float& rec);
-	virtual Point3F hit_normal(const Ray& ray, float t) const;
+	virtual bool hit(const Ray& ray, double t0, double t1, double& rec);
+	virtual Point3F hit_normal(const Ray& ray, double t) const;
 	Vertice center() const { return center_; }
 	float radius() const { return radius_; }
 
-	virtual Point3F change_normal(const Ray& ray, float t) const;
+	virtual Point3F change_normal(const Ray& ray, double t) const;
 };
